+++ lib/BibSpace/Controller/Publications.pm
@@ -71,8 +71,8 @@ sub all_recently_added {
 
   my @all = Fget_publications_main_hashed_args($self, {year => undef});
   my @added_entries = sort { $b->creation_time cmp $a->creation_time } @all;
-  my $num_found = scalar @added_entries;
-  my $limit = $num < $num_found ? $num : $num_found;
+  my $num_found     = scalar @added_entries;
+  my $limit         = $num < $num_found ? $num : $num_found;
   @added_entries = @added_entries[0 .. $limit - 1];
 
   my @filtered = Fget_publications_main_hashed_args($self, {}, \@added_entries);
@@ -92,8 +92,8 @@ sub all_recently_modified {
 
   my @all = Fget_publications_main_hashed_args($self, {year => undef});
   my @modified_entries = sort { $b->modified_time cmp $a->modified_time } @all;
-  my $num_found = scalar @modified_entries;
-  my $limit = $num < $num_found ? $num : $num_found;
+  my $num_found        = scalar @modified_entries;
+  my $limit            = $num < $num_found ? $num : $num_found;
   @modified_entries = @modified_entries[0 .. $limit - 1];
 
   my @filtered