+++ lib/BibSpace/Functions/BackupFunctions.pm
@@ -4,7 +4,10 @@ use BibSpace::Functions::MySqlBackupFunctions;
 use BibSpace::Model::Backup;
 use BibSpace::Functions::Core;
 use BibSpace::Functions::FDB;
+use BibSpace::Backend::SmartBackendHelper;
 
+use JSON -convert_blessed_universally;
+use BibSpace::Model::SerializableBase::BibSpaceDTO;
 use Storable;
 
 use Data::Dumper;
@@ -28,7 +31,9 @@ our @EXPORT = qw(
   find_backup
   read_backups
   do_storable_backup
+  do_json_backup
   do_mysql_backup
+  restore_json_backup
   restore_storable_backup
   delete_old_backups
 );
@@ -40,7 +45,7 @@ sub find_backup {
   my $dir  = shift;
 
   my @backup_files
-    = Path::Tiny->new($dir)->children(qr/^backup_$uuid.*\.(dat|sql)$/);
+    = Path::Tiny->new($dir)->children(qr/^backup_$uuid.*\.(dat|sql|json)$/);
   return if scalar(@backup_files) == 0;
 
   my $file   = shift @backup_files;
@@ -48,13 +53,13 @@ sub find_backup {
   $backup->dir($dir);
   return $backup;
 }
-## Trivial DAO ALL
 
+## Trivial DAO ALL
 sub read_backups {
   my $dir = shift;
 
   my @backup_files
-    = Path::Tiny->new($dir)->children(qr/^backup_.*\.(dat|sql)$/);
+    = Path::Tiny->new($dir)->children(qr/^backup_.*\.(dat|sql|json)$/);
   my @backups;
   foreach my $file (@backup_files) {
     my $backup = Backup->parse($file->basename);
@@ -62,6 +67,24 @@ sub read_backups {
     push @backups, $backup;
   }
   return @backups;
+}
+
+sub do_json_backup {
+  my $app = shift;
+  my $name = shift // 'normal';
+
+  my $backup_dir = Path::Tiny->new($app->get_backups_dir)->relative;
+  $backup_dir =~ s!/*$!/!;
+
+  my $backup = Backup->create($name, "json");
+  $backup->dir("" . $backup_dir);
+  my $path = "" . $backup->get_path;
+
+  my $dtoObject  = BibSpaceDTO->fromLayeredRepo($app->repo);
+  my $jsonString = $dtoObject->toJSON;
+  path($backup->get_path)->spew($jsonString);
+
+  return $backup;
 }
 
 sub do_storable_backup {
@@ -95,6 +118,77 @@ sub do_mysql_backup {
   dump_mysql_to_file($backup->get_path, $app->config);
 
   return $backup;
+}
+
+sub restore_json_backup {
+  my $backup = shift;
+  my $app    = shift;
+
+  my $jsonString = '{}';
+  my $file       = path($backup->get_path);
+  if (not $file->exists or not $file->is_file) {
+    $app->logger->warn("Cannot restore JSON backup from file "
+        . $file
+        . " - file does not exist.");
+    return;
+  }
+  $jsonString = $file->slurp_utf8;
+
+  my $success;
+  my $dto = BibSpaceDTO->new();
+  my $decodedDTO;
+  try {
+    $decodedDTO = $dto->toLayeredRepo(
+      $jsonString,
+      $app->repo->lr->uidProvider,
+      $app->repo->lr->preferences
+    );
+    $success = 1;
+  }
+  catch {
+    $app->logger->error("Exception during JSON restore: $_");
+    $success = undef;
+  };
+
+  # First Models, then Relations
+  if (defined $success and $success == 1){
+    my @layers = $app->repo->lr->get_all_layers;
+    foreach (@layers) { $_->reset_data }
+    # $app->repo->lr->reset_uid_providers;
+
+    for my $type (@{$app->repo->entities}, @{$app->repo->relations}) {
+      my $arrayRef = $decodedDTO->data->{$type};
+      # Authors reference each other, so the order of restoring is important
+      # First masters, because they always reference themselves, and then minions
+      my @waitingLine;
+      for my $object (@$arrayRef) {
+        try {
+          if ($type eq "Author" && $object->is_minion){
+            print "Putting $object->{uid} to waiting line\n";
+            push @waitingLine, $object;
+          }
+          else {
+            $app->repo->lr->save($type, $object);
+          }
+        }
+        catch {
+           $app->logger->warn("Skipped restoring object of type '$type' from JSON backup. Error: $_");
+        };
+      }
+      for my $object (@waitingLine) {
+        try {
+          print "Adding $object->{uid} from waiting line\n";
+          $app->repo->lr->save($type, $object);
+        }
+        catch {
+           $app->logger->warn("Could not restore minion Author from waitingLine. Error: $_");
+        };
+      }
+    }
+    BibSpace::Backend::SmartBackendHelper::linkData($app);
+    # TODO: Do linking in the smart layer!
+  }
+  return $success;
 }
 
 sub restore_storable_backup {